Move preallocate fallback into native
When preallocating a file, some systems (Windows) don't have a native
method to do so. However, using setLength in Java causes Windows to
fully allocate the file. The blobcache's preallocate method already
falls back to using setLength. But the preallocation tests in native
access don't have that fallback. This commit moves the fallback
implementation into native access so that preallocate tests can operate
the expect the same on all platforms.