Make transport resources wrapper available in tests
The transport resources build service encapsulates accessing transport
resources but it requires a gradle build to setup. Yet in tests it can
be useful to use this same access to check the state of the resources.
This commit moves the bulk of the service to a TransportVersionResources
class that can be constructed either in a gradle context as a build
service, or in the groovy functional tests using the fake build dir.